+++ crates/cxx-qt-lib/src/gui/qvector4d.rs
@@ -133,8 +133,12 @@ mod ffi {
     }
 }
 
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 /// The QVector4D class represents a vector or vertex in 4D space.
 #[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[repr(C)]
 pub struct QVector4D {
     v: [f32; 4],